Move stuff over to its own crate
