+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
 	"client/consensus/common",
 	"client/consensus/relay-chain",
 	"client/network",
+	"client/pov-recovery",
 	"client/service",
 	"pallets/dmp-queue",
 	"pallets/parachain-system",
